+++ api/src/main/resources/application.properties
@@ -161,6 +161,8 @@ org.terrakube.api.module.cache.schedule=${ModuleCacheSchedule:0 */3 * ? * *}
 #######################
 # Dynamic Credentials #
 #######################
+org.terrakube.dynamic.credentials.kid=${DynamicCredentialId:03446895-220d-47e1-9564-4eeaa3691b42}
+org.terrakube.dynamic.credentials.ttl=${DynamicCredentialTtl:30}
 org.terrakube.dynamic.credentials.public-key-path=${DynamicCredentialPublicKeyPath:}
 org.terrakube.dynamic.credentials.private-key-path=${DynamicCredentialPrivateKeyPath:}
